MNT: refactor the adapter code for image sequences (jpeg and tiff)
